requirements.txt
pandas
requests
dnspython
matplotlib
streamlit
plotly
